backend/requirements.txt
# Create a virtual environment
# for python
#python -m venv venv
#source venv/bin/activate

# for windows
#>> python -m venv ./venv
# Activate the virtual environment
#>> .\venv\Scripts\activate

# install dependicies as
#>> pip install -r requirement.txt

# git repository
#>> git add .
#>> git commit -m "commit message"
#>> git push -u origin main

# steup files
#>> python template.py

# Install required dependencies
Flask
Flask-JWT-Extended
Flask-Bcrypt
Flask-SQLAlchemy
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
setuptools
Flask-Cors
pillow
pytesseract
google-generativeai
python-dateutil
mysql-connector-python
joblib
pandas
scikit-learn